add find details of medicine feature in, still requires styling
